fix: resolve CI clippy warnings and benchmark workflow failures
- Add #[allow(clippy::derive_ord_xor_partial_ord)] on PdfPoints struct
- Remove redundant `as usize` casts in font.rs
- Fix cache-benchmark-harness $(uname -m) not expanding in YAML with: context
- Reduce WASM benchmark timeout (120s) and iterations (1) to fit job limit
- Make aggregate job resilient to missing artifacts from cancelled upstream jobs

+++ crates/kreuzberg-pdfium-render/src/pdf/font.rs
@@ -758,7 +758,7 @@ impl<'a> PdfFont<'a> {
             return String::new();
         }
 
-        let mut buffer = create_byte_buffer(buffer_length as usize);
+        let mut buffer = create_byte_buffer(buffer_length);
 
         let result =
             self.bindings
@@ -794,7 +794,7 @@ impl<'a> PdfFont<'a> {
             return String::new();
         }
 
-        let mut buffer = create_byte_buffer(buffer_length as usize);
+        let mut buffer = create_byte_buffer(buffer_length);
 
         let result =
             self.bindings
